Adaptando a nueva organizacion

+++ include/Utils/Data.hpp
@@ -12,7 +12,7 @@ class Data{
         static const int MAXCHAR = 1000;
         const char* filename;
         int size;
-        vector<vector<float>> X;
+        vector<vector<double>> X;
         vector<vector<int>> Y;
         vector<int> Yneg;
         vector<int> Ypos;
