+++ src/Utils/Data.cpp
@@ -31,7 +31,7 @@ void Data::init(const char* filename,const int tamXi,const int tamYi){
     this->tamYi=tamYi;
 
     this->size=countLinesFile();
-    this->X.resize(this->size,vector<float>(tamXi));
+    this->X.resize(this->size,vector<double>(tamXi));
     this->Y.resize(this->size,vector<int>(tamYi));
     this->Yneg.resize(tamYi);
     this->Ypos.resize(tamYi);
